pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "value-screener"
version = "0.1.0"
description = "A pragmatic value stock screener with EV/EBIT, FCF yield, ROIC filters and more."
requires-python = ">=3.9"
authors = [{name = "Your Name"}]
readme = "README.md"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "yfinance>=0.2.43",
  "click>=8.1",
  "PyYAML>=6.0",
  "tabulate>=0.9",
  "tqdm>=4.66",
  "pydantic>=2.7"
]

[project.scripts]
value-screener = "value_screener.cli:main"
